Bullet
A bullet is a kinetic projectile and the component of firearm ammunition that is expelled from the gun barrel during shooting. The term is from Middle French and originated as the diminutive of the word boulle (boullet), which means "small ball".[1] Bullets are made of a variety of materials such as copper, lead, steel, polymer, rubber and even wax. They are available either singly as in muzzleloading and cap and ball firearms[2] or as a component of paper cartridges[3], but much more commonly in the form of metallic cartridges.[4] Bullets are made in a large number of shapes and constructions depending on the intended applications, including specialized functions such as hunting, target shooting, training and combat.

Though the word "bullet" is often used incorrectly in colloquial language to refer to a cartridge round, a bullet is not a cartridge but rather a component of one.[5] A round of ammunition cartridge is a combination package of the bullet (which is the projectile),[6] the case (which holds everything together), the propellant (which provide majority of the energy to launch the projectile) and the primer (which ignites the propellant). This use of the term "bullet" when intending to describe a cartridge often leads to confusion when the components of a cartridge are specifically referred to.

Bullet sizes are expressed by their weights and diameters (referred to as "calibers") in both imperial[7] and metric measurement systems. For example: 55 grain .223 caliber bullets are of the same weight and caliber as 3.56 gram 5.56mm caliber bullets.

The bullets used in many cartridges are fired at muzzle velocities faster than the speed of sound[8][9] — about 343 metres per second (1,130 ft/s) in dry air at 20 °C (68 °F) — and thus can travel a substantial distance to a target before a nearby observer hears the sound of the shot. The sound of gunfire (i.e. the muzzle report) is often accompanied with a loud bullwhip-like crack as the supersonic bullet pierces through the air creating a sonic boom. Bullet speeds at various stages of flight depend on intrinsic factors such as its sectional density, aerodynamic profile and ballistic coefficient, and extrinsic factors such as barometric pressure, humidity, air temperature and wind speed.[10][11] Subsonic cartridges fire bullets slower than the speed of sound so there is no sonic boom. This means that a subsonic cartridge, such as .45 ACP, can be substantially quieter than a supersonic cartridge such as the .223 Remington, even without the use of a suppressor.[12]

Bullets do not normally contain explosives,[13] but damage the intended target by transferring kinetic energy upon impact and penetration (see terminal ballistics).